refactor: dedup id derivation and index writes
A simplification pass on the branch turned up a few copy-pasted blocks worth folding

Page ids and callable-result ids now share one `NodeId::from_seed`, so the sha256-to-ULID derivation lives in one place (outl-actions drops its sha2 dep). The two atomic index saves share a `write_atomic` helper for the tmp, fsync, rename-with-cleanup skeleton. The two catch-up test wrappers collapse into one taking an `Option`. Also drops four dead crate-root re-exports and a duplicated comment. Behavior-preserving, all tests green
