+++ crates/outl-actions/src/lib.rs
@@ -117,8 +117,7 @@ pub use person::{search_persons, PERSON_TYPE, TYPE_KEY};
 pub use resolve::{open_or_create_by_name, open_or_create_by_ref};
 pub use sync::{FileSyncTransport, OpsFileSnapshot, PeerHealthSnapshot, SyncEngine, SyncTransport};
 pub use template::{
-    call_target_name, inject_call_params, instantiate_template, list_templates,
-    parse_call_invocation, parse_call_params, resolve_call, run_callable_block, CallResolution,
+    instantiate_template, list_templates, parse_call_invocation, resolve_call, run_callable_block,
     TemplateEntry, FROM_TEMPLATE_KEY, JOURNAL_TEMPLATE_NAME, PARAMS_KEY, TEMPLATE_KEY,
 };
 pub use todo::{cycle_todo, split_todo, TodoState, DONE_PREFIX, TODO_PREFIX};
